# docs/openapi.json
{
  "openapi": "3.0.3",
  "paths": {
...
    "/tickets": {
      "post": {
        "operationId": "TicketsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ticket created."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a ticket for the authenticated requester",
        "tags": [
          "tickets"
        ]
      },
      "get": {
        "operationId": "TicketsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "OPEN",
                "IN_PROGRESS",
                "WAITING_USER",
                "RESOLVED",
                "CLOSED"
              ]
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "LOW",
                "MEDIUM",
                "HIGH",
                "CRITICAL"
              ]
            }
          },
          {
            "name": "assigneeId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated tickets."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List visible tickets with pagination, search and filters",
        "tags": [
          "tickets"
        ]
      }
    },
    "/tickets/{id}": {
      "get": {
        "operationId": "TicketsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a visible ticket",
        "tags": [
          "tickets"
        ]
      },
      "patch": {
        "operationId": "TicketsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update workflow, priority or assignment",
        "tags": [
          "tickets"
        ]
      }
    },
...
    "/reports/tickets": {
      "post": {
        "operationId": "ReportsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketReportDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Report queued for asynchronous generation."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Queue a ticket PDF report",
        "tags": [
          "reports"
        ]
      }
    },
    "/reports/{id}": {
      "get": {
        "operationId": "ReportsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read the status of an owned report job",
        "tags": [
          "reports"
        ]
      }
    },
    "/reports/{id}/content": {
      "get": {
        "operationId": "ReportsController_download",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Download a completed ticket report",
        "tags": [
          "reports"
        ]
      }
    },
    "/notifications": {
      "get": {
        "operationId": "NotificationsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "unread",
            "required": false,
            "in": "query",
            "description": "When true, return only unread notifications.",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List notifications for the authenticated user",
        "tags": [
          "notifications"
        ]
      }
    },
    "/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationsController_markRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark an owned notification as read",
        "tags": [
          "notifications"
        ]
      }
    },
    "/admin/audit-events": {
      "get": {
        "operationId": "GovernanceController_listAuditEvents",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 50,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "action",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resourceType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resourceId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actorId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List append-only audit events",
        "tags": [
          "governance"
        ]
      }
    },
    "/admin/users/{id}/roles": {
      "patch": {
        "operationId": "GovernanceController_updateRoles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRolesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Replace a user role set without removing the final admin",
        "tags": [
          "governance"
        ]
      }
    },
    "/admin/job-failures": {
      "get": {
        "operationId": "GovernanceController_listJobFailures",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 50,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List dead-letter and reprocessed jobs",
        "tags": [
          "governance"
        ]
      }
    },
    "/admin/job-failures/{id}/reprocess": {
      "post": {
        "operationId": "GovernanceController_reprocessJob",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reprocess an allowlisted dead-letter job once",
        "tags": [
          "governance"
        ]
      }
    },
...
  },
  "info": {
    "title": "Secure Service Desk API",
    "description": "Secure, modular service desk API for portfolio demonstration. The MVP is single-tenant and manages identity, tickets, files, audit, notifications and reports in code.",
    "version": "0.1.0",
    "contact": {},
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/license/mit/"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Local demo reverse proxy"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Standalone API process"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "service_desk_refresh"
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-csrf-token",
        "description": "Double-submit CSRF token returned by register or login."
      }
    },
    "schemas": {
...
      "CreateTicketDto": {
        "type": "object",
        "properties": {
          "subject": {
            "type": "string",
            "example": "Cannot access the customer portal",
            "maxLength": 160
          },
          "description": {
            "type": "string",
            "example": "The portal returns an unauthorized response after signing in.",
            "maxLength": 10000
          }
        },
        "required": [
          "subject",
          "description"
        ]
      },
      "Object": {
        "type": "object",
        "properties": {}
      },
      "UpdateTicketDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "description": "Current ticket version used for optimistic concurrency control.",
            "minimum": 0
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "WAITING_USER",
              "RESOLVED",
              "CLOSED"
            ]
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "CRITICAL"
            ]
          },
          "assigneeId": {
            "type": "object",
            "description": "MongoDB user ID. Send null to remove the assignee.",
            "nullable": true
          },
          "resolution": {
            "type": "string",
            "description": "Required when resolving a ticket.",
            "minLength": 5,
            "maxLength": 5000
          }
        },
        "required": [
          "version"
        ]
      },
...
      "CreateTicketReportDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "WAITING_USER",
              "RESOLVED",
              "CLOSED"
            ]
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "CRITICAL"
            ]
          },
          "maxRows": {
            "type": "object",
            "default": 500,
            "minimum": 1,
            "maximum": 2000
          }
        }
      },
      "UpdateUserRolesDto": {
        "type": "object",
        "properties": {
          "roles": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "string",
              "enum": [
                "ADMIN",
                "SUPPORT",
                "USER"
              ]
            }
          }
        },
        "required": [
          "roles"
        ]
      },
...
    }
  }
}